tests/baseline/protocol-1.2/http/tests/inspector/resource-tree/cached-resource-metadata-expected.txt

CONSOLE MESSAGE: line 1: Done.
Verify that cached resource has metadata.

Last modified: null
Content size: null
